Compile MIDL output C file into client and servers to replace hack
Previously, instead of compiling in the C implementation of the interface
output by the MIDL compiler, the MIDL included manual C++ output that
defined a few of the necessary constants. This commit removes that manual
hack, and replaces it by actually compiling in the interface C file, not
just the header file.