Change ExecutablePingServer registration to register a type library
Upon registration, in addition to registering a class id and app id, the
ExecutablePingServer will now also register a type library to the TypeLib
table, as well as registering IPingable to the Interface table and adding
a TypeLib key to the class id registration (both necessary to support the
registration of a type library).

Notably, it also specifies that IPingable with use the Automation
marshaller (by setting the ProxyStubClsid32 to the UUID of the Automation
marshaller), which is a break from previous implementation, which used the
custom PingProxyStubDll.dll for marshalling. This will allow the
ExecutablePingServer to operate without registering any custom marshalling
code (which seems to be necessary for the eventual goal of compiling the
project with gcc).