Change type library path hack to be execution directory independent
The previous hack for registering the location of the type library when
registering the ExecutablePingServer relied on running the executable from
a particular directory. This new hack will be independent of the directory
you run it from...but will require hard coding the project directory into
a source file within the project.