Fix component reference counting issue in ExecutablePingServer
Despite the fact that the CoPingEngineFactory was having its lifetime
managed manually, it still incremented the component level reference
count. Because the component destroyed the CoPingEngineFactory upon exit
the server were experience deadlock, with the factory not being destroyed
until the server to exited, but the server not exiting until the factory
had been destroyed. This commit fixes that issue by no longer incrementing
or decrementing the component level reference count in the factory code.

In addition, in order to keep the server output on screen for a longer
period of time (whereas before this issue was fixed, it was up
indefinitely) the pre-exit sleep has been raised from 5 seconds to 60
seconds.